lfec/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
feedparser>=6.0.10
Pillow>=10.0.0
